requirements.txt
streamlit
pandas
matplotlib
pdfplumber
numpy
xlsxwriterselenium==4.34.0
